pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "PySciVisToolKit"
authors = [
    {name = "Kaiyuan Tang", email = "dullpigeon@gmail.com"},
]
description = "a simple toolkit for scivis"
readme = "README.md"
requires-python = ">=3.8.0"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "icecream==2.1.3",
    "imageio==2.22.2",
    "lpips==0.1.4",
    "netCDF4==1.7.2",
    "numpy==1.23.3",
    "opencv_python==4.6.0.66",
    "opencv_python_headless==4.9.0.80",
    "Pillow==11.0.0",
    "pytorch_msssim==0.2.1",
    "PyYAML==6.0.2",
    "scikit_image==0.19.3",
    "scikit_learn==1.2.1",
    "scipy==1.10.0",
    "tifffile==2024.7.21",
    "torch==2.0.0+cu118",
    "torchvision==0.15.1+cu118",
    "tqdm==4.64.1",
    "tyro==0.9.2",
    "Wand==0.6.11",
]
version = "0.0.1"


[project.scripts]
diff_img = "src.diff_img:entrypoint"
img_lpips = "src.img_lpips:entrypoint"
img_psnr = "src.img_psnr:entrypoint"
img_ssim = "src.img_ssim:entrypoint"
vol_psnr = "src.vol_psnr:entrypoint"
norm = "src.norm:entrypoint"
trim = "src.trim:entrypoint"
nc2raw = "src.nc2raw:entrypoint"
tiff2vol = "src.tiff2vol:entrypoint"
imgs2video = "src.imgs2video:entrypoint"

[tool.setuptools.packages.find]
include = ["src*"]